app/requirements.txt
# Language detection
langdetect==1.0.9
aiohappyeyeballs==2.4.3 ; python_version >= "3.12" and python_version < "4.0"
aiohttp==3.10.10 ; python_version >= "3.12" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
anyio==4.6.2.post1 ; python_version >= "3.12" and python_version < "4.0"
attrs==24.2.0 ; python_version >= "3.12" and python_version < "4.0"
certifi==2024.8.30 ; python_version >= "3.12" and python_version < "4.0"
charset-normalizer==3.4.0 ; python_version >= "3.12" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and platform_system == "Windows"
dataclasses-json==0.6.7 ; python_version >= "3.12" and python_version < "4.0"
distro==1.9.0 ; python_version >= "3.12" and python_version < "4.0"
fastapi==0.115.2 ; python_version >= "3.12" and python_version < "4.0"
filelock==3.16.1 ; python_version >= "3.12" and python_version < "4.0"
frozenlist==1.4.1 ; python_version >= "3.12" and python_version < "4.0"
fsspec==2024.12.0 ; python_version >= "3.12" and python_version < "4.0"
greenlet==3.1.1 ; python_version < "3.13" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version >= "3.12"
h11==0.14.0 ; python_version >= "3.12" and python_version < "4.0"
httpcore==1.0.6 ; python_version >= "3.12" and python_version < "4.0"
httpx==0.27.2 ; python_version >= "3.12" and python_version < "4.0"
huggingface-hub==0.27.1 ; python_version >= "3.12" and python_version < "4.0"
idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
jinja2==3.1.4 ; python_version >= "3.12" and python_version < "4.0"
jiter==0.6.1 ; python_version >= "3.12" and python_version < "4.0"
jsonpatch==1.33 ; python_version >= "3.12" and python_version < "4.0"
jsonpointer==3.0.0 ; python_version >= "3.12" and python_version < "4.0"
langchain-community
langchain-core
langchain-experimental
langchain-openai
langchain-text-splitters
langchain-qdrant
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.0
langchain==0.3.3
langsmith==0.1.135
markupsafe==3.0.2
marshmallow==3.23.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai>=1.52.0 
orjson==3.10.7
packaging==24.1
pandas==2.2.3
propcache==0.2.0
psycopg2-binary==2.9.10
psycopg2==2.9.10
pydantic-core==2.23.4
pydantic-settings==2.6.0
pydantic==2.9.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyyaml==6.0.2
regex==2024.9.11
requests-toolbelt==1.0.0
requests==2.32.3
safetensors==0.5.2
six==1.16.0
sniffio==1.3.1
sqlalchemy==2.0.36
starlette==0.40.0
tenacity==8.5.0
tiktoken==0.8.0
tokenizers==0.21.0
tqdm==4.66.5
transformers==4.48.0
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
urllib3<2.0.0,>=1.26.14
uvicorn==0.32.0
yarl==1.15.4
qdrant-client==1.13.2
fastembed==0.3.6
torch==2.2.0
agno==1.7.3
pybreaker>=0.8.0

# Recipe Search Dependencies
sentence-transformers==3.2.0
scikit-learn==1.5.2
openpyxl==3.1.2

langchain-groq
groq

# JWT Authentication
python-jose[cryptography]==3.3.0

# Azure AD SSO
msal>=1.24.0
cachetools>=5.3.0

# Document Processing
# unstructured[all-docs] includes OCR support for images and screenshots
# 
# REQUIRED SYSTEM DEPENDENCIES (install BEFORE pip install):
#   - macOS: 
#       brew install tesseract libmagic
#   - Ubuntu/Debian: 
#       sudo apt-get install -y libmagic1 tesseract-ocr
#   - Windows: 
#       Use python-magic-bin instead (no system deps needed)
#       OR download tesseract from https://github.com/UB-Mannheim/tesseract/wiki
#
unstructured[all-docs]==0.18.18
python-magic>=0.4.27; sys_platform != 'win32'
python-magic-bin>=0.4.14; sys_platform == 'win32'
pillow>=10.4.0
pypdf>=5.1.0
python-docx>=1.1.2
python-pptx>=1.0.2
